Localhost uses psutil instead of SSHing to the local IP to get cluster information. Moved cluster information to utils
