+++ benchmark/config.py
@@ -39,7 +39,7 @@ STEPS_PER_EPOCH = 500
 MOST_MEMORY_MULTIPLIER = 5
 MAX_READ_THREADS_TRAINING = 32
 
-DEFAULT_HOSTS = "127.0.0.1"
+DEFAULT_HOSTS = ["127.0.0.1",]
 
 MPI_RUN_BIN = os.environ.get("MPI_RUN_BIN", MPIRUN)
 ALLOW_RUN_AS_ROOT = True
